# level2/task2/docs/swagger.yaml
swagger: "2.0"
basePath: /api/v1
definitions:
...
  controller.BallotVoteRequest:
    properties:
      proposal_id:
        description: 提案ID
        type: string
    required:
    - proposal_id
    type: object
  controller.BlockResponse:
    properties:
      hash:
        example: 0x1234567890abcdef1234567890abcdef1234567890abcdef1234567890abcdef
        type: string
      number:
        example: "123456"
        type: string
      parentHash:
        example: 0xabcdef1234567890abcdef1234567890abcdef1234567890abcdef1234567890
        type: string
      timestamp:
        example: 1634567890
        type: integer
      txCount:
        example: 10
        type: integer
    type: object
...
  controller.ErrorResponse:
    properties:
      code:
        example: 500
        type: integer
      message:
        example: 服务器内部错误
        type: string
    type: object
...
  controller.SuccessResponse:
    properties:
      code:
        example: 200
        type: integer
      data: {}
      errMsg:
        example: ""
        type: string
    type: object
  controller.TransactionRequest:
    properties:
      amount:
        description: 十六进制字符串表示的wei金额，例如"0xde0b6b3a7640000"表示1 ETH
        example: "0xde0b6b3a7640000"
        type: string
      from:
        example: 0x123...
        type: string
      to:
        example: 0x456...
        type: string
    required:
    - amount
    - from
    - to
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: 提供以太坊区块链相关的API服务，包括账户管理、交易处理、合约部署等功能
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Web3 区块链接口服务
  version: "1.0"
paths:
...
  /block/{number}:
    get:
      consumes:
      - application/json
      description: 根据区块编号获取区块信息
      parameters:
      - description: 区块编号
        in: path
        name: number
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.BlockResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: 获取区块
      tags:
      - 区块
  /block/hash/{hash}:
    get:
      consumes:
      - application/json
      description: 根据区块哈希获取区块信息
      parameters:
      - description: 区块哈希
        in: path
        name: hash
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.BlockResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: 根据区块哈希获取区块信息
      tags:
      - 区块
  /block/info:
    get:
      consumes:
      - application/json
      description: 根据区块编号获取区块信息
      parameters:
      - description: 区块编号
        in: query
        name: number
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.BlockResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: 获取区块信息
      tags:
      - 区块
  /block/latest:
    get:
      consumes:
      - application/json
      description: 获取最新区块的信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.BlockResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: 获取最新区块
      tags:
      - 区块
  /block/subscribe:
    get:
      consumes:
      - application/json
      description: 使用WebSocket订阅新区块
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.BlockResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: 订阅区块
      tags:
      - 区块
    post:
      consumes:
      - application/json
      description: 启动后台服务订阅新区块
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: 启动区块订阅
      tags:
      - 区块
...
  /contracts/Ballot/vote:
    post:
      consumes:
      - application/json
      description: 调用Ballot合约的vote方法进行投票
      parameters:
      - description: 投票请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.BallotVoteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: 投票功能
      tags:
      - Ballot
  /contracts/Ballot/winner:
    get:
      consumes:
      - application/json
      description: 调用Ballot合约的winningProposal方法获取当前得票最多的提案ID
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.SuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: 获取获胜提案ID
      tags:
      - Ballot
  /contracts/Ballot/winner-name:
    get:
      consumes:
      - application/json
      description: 调用Ballot合约的winnerName方法获取当前得票最多的提案名称
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.SuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: 获取获胜提案名称
      tags:
      - Ballot
...
  /transaction/create:
    post:
      consumes:
      - application/json
      description: 创建一笔交易，并发送给以太坊网络
      parameters:
      - description: 交易请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.TransactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 返回创建交易的结果
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: 创建交易失败
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: 创建交易
      tags:
      - 交易
  /transaction/raw:
    post:
      consumes:
      - application/json
      description: 创建一笔原始交易，并发送给以太坊网络
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: 创建原始交易
      tags:
      - 交易
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey